Refactoring and smaller fixes.
- Created VS2019 projects.
- Used some useful modern C++ features.
- Changed integer types used as function parameters to match OpenGL calls.
- Coresponding GL state tracking variables are reset when a resource is deleted.

+++ include/GLSlayer/VertexFormat.h
@@ -25,7 +25,7 @@ namespace gls
 		DataType type;
 		bool integer;
 		bool normalized;
-		size_t offset;
+		uint offset;
 		uint divisor;
 	};
